Issue #100 - Add index helpers to helpers modules
