+++ helpers.js
@@ -153,6 +153,87 @@ function filterWebCompatSeeAlso(bug) {
     return !seeAlsoArray.some(url => webCompatRegexp.test(url));
 }
 
+/**
+ * Return the End of the week the specified date belongs to.
+ * @param {Date} date the date for which to
+ * @returns the date of the end of the week
+ */
+function getEOW(date) {
+    const param = new Date(date);
+    param.setDate(param.getDate() - param.getDay() + 7);
+    return new Date(param - 1);
+}
+
+/**
+ * Return the list of query dates for a given inputDate
+ * @param {Date} inputDate the date to start with.
+ * @returns an Array with all dates to gather bugs for
+ */
+function getQueryDates(inputDate) {
+    const queryDates = [];
+    if (inputDate) {
+        // We want to consider open bugs only until the end of the given week.
+        const parsed = new Date(inputDate);
+        const today = new Date();
+        if (isNaN(parsed)) {
+            throw new Error("Wrong maxDate format: use yyyy-mm-dd");
+        }
+
+        if (!inputDate.includes("-")) {
+            // An entire year is specified.
+            for (let i = 0; i < 52; i++) {
+                queryDates.push(getEOW(parsed));
+                parsed.setDate(parsed.getDate() + 7);
+                if (getEOW(parsed) > today) {
+                    // Stop if we get into future dates (the Tranco list won't
+                    // have anything useful for us).
+                    break;
+                }
+            }
+        } else if (inputDate.indexOf("-") === inputDate.lastIndexOf("-")) {
+            // An entire month is specified.
+            const month = getEOW(parsed).getMonth();
+            for (let i = 0; i < 5; i++) {
+                queryDates.push(getEOW(parsed));
+                parsed.setDate(parsed.getDate() + 7);
+                if (getEOW(parsed).getMonth() !== month || getEOW(parsed) > today) {
+                    // Stop if the fifth consecutive Sunday falls into the next
+                    // month, or we get into future dates.
+                    break;
+                }
+            }
+        } else {
+            // A single date is specified.
+            queryDates.push(getEOW(parsed));
+        }
+    } else {
+        queryDates.push(inputDate);
+    }
+
+    return queryDates;
+}
+
+/**
+ * Return the list of query dates until the present, starting
+ * at the specified date.
+ * @param {Date} inputDate the date to resume with
+ * @returns an Array with all dates to gather bugs for
+ */
+function resumeQueryDates(inputDate) {
+    const queryDates = [];
+    const parsed = new Date(inputDate);
+    const today = new Date();
+    if (isNaN(parsed)) {
+        throw new Error("Wrong maxDate format: use yyyy-mm-dd");
+    }
+    while(getEOW(parsed) < today) {
+        queryDates.push(getEOW(parsed));
+        parsed.setDate(parsed.getDate() + 7);
+    }
+
+    return queryDates;
+}
+
 module.exports = {
     bugzillaRetry,
     formatDateForAPIQueries,
@@ -162,8 +243,11 @@ module.exports = {
     getBugzillaProducts,
     getBugzillaStatuses,
     getBzLink,
+    getEOW,
+    getQueryDates,
     filterWebCompatSeeAlso,
     isMobile,
     isDesktop,
     isNotQA,
+    resumeQueryDates,
 }
